[Refactoring] Seperate SSDFileHandler class
