+++ TeamProject_SSD/TeamProject_SSD.vcxproj
@@ -130,11 +130,13 @@
     <ClCompile Include="main.cpp" />
     <ClCompile Include="ShellTestApp.cpp" />
     <ClCompile Include="ssd.cpp" />
+    <ClCompile Include="SSDFileHandler.cpp" />
     <ClCompile Include="testscript.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="ShellTestApp.h" />
     <ClInclude Include="ssd.h" />
+    <ClInclude Include="SSDFileHandler.h" />
     <ClInclude Include="testscript.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
