+++ TeamProject_SSD/ssd.cpp
@@ -1,48 +1,11 @@
 #include "ssd.h"
-#include <io.h>
-#include <string>
-
-void SSDFileHandler::ReadFromNANDFile(unordered_map<unsigned int, unsigned int>& umDataSet)
-{
-	string sIndex, sValue;
-
-	if (IsNANDFileExist())
-	{
-		ifstream fin(sNandFileName);
-		while (!fin.eof())
-		{
-			fin >> sIndex >> sValue;
-			umDataSet.insert({ stoi(sIndex), stoul(sValue, nullptr, 16) });
-		}
-	}
-}
-
-void SSDFileHandler::WriteToNANDFile(std::unordered_map<unsigned int, unsigned int>& umDataSet)
-{
-	ofstream fout(sNandFileName);
-
-	for (const auto& pair : umDataSet) 
-	{
-		fout << dec << pair.first << " " << "0x" << uppercase << hex << setw(8) << setfill('0') << pair.second << endl;
-	}
-}
-
-void SSDFileHandler::WriteHexReadValueToResultFile(unsigned int nValue)
-{
-	ofstream fResultFile(sResultFileName);
-	fResultFile << "0x" << uppercase << hex << setw(8) << setfill('0') << nValue;
-}
-
-bool SSDFileHandler::IsNANDFileExist()
-{
-	return _access(&*sNandFileName.begin(), 0) == 0;
-}
 
 unsigned int SSD::Read(unsigned int nLBA)
 {
 	ValidateParameter(nLBA);
 
 	unsigned int nReadValue = DEFAULT_READ_VALUE;
+	
 	unordered_map<unsigned int, unsigned int> umDataSet;
 	ssdFileHandler.ReadFromNANDFile(umDataSet);
